moving semvercmp to separate file
