+++ lib/transport/chrome-extension.js
@@ -2,6 +2,7 @@
 
 var Promise = require('promise');
 var ChromeMessages = require('../chrome-messages');
+var semvercmp = require('./semvercmp');
 
 var EXTENSION_ID = 'jcjjhjgimijdkoamemaghajlhegmoclj';
