+++ lib/transport/http.js
@@ -2,6 +2,7 @@
 
 var extend = require('extend');
 var http = require('../http');
+var semvercmp = require('./semvercmp');
 
 var DEFAULT_URL = 'https://localback.net:21324';
 
@@ -139,21 +140,5 @@ HttpTransport.prototype.callSync = function (sessionId, type, message) {
     return this._requestSync(callOptions(sessionId, type, message));
 };
 
-// Taken from https://github.com/substack/semver-compare/blob/master/index.js
-function semvercmp(a, b) {
-    var pa = a.split('.');
-    var pb = b.split('.');
-
-    for (var i = 0; i < 3; i++) {
-        var na = Number(pa[i]);
-        var nb = Number(pb[i]);
-        if (na > nb) return 1;
-        if (nb > na) return -1;
-        if (!isNaN(na) && isNaN(nb)) return 1;
-        if (isNaN(na) && !isNaN(nb)) return -1;
-    }
-
-    return 0;
-}
 
 module.exports = HttpTransport;
